Rename store.js to index.js
This is to match the new vue-cli defaults
